adding deletion validation if product in cart

+++ senior-test-artur/src/main/java/com/senior/services/ProductServiceImpl.java
@@ -1,11 +1,14 @@
 package com.senior.services;
 
+import com.senior.controllers.CartProductController;
 import com.senior.dto.request.CartRequest;
 import com.senior.dto.request.ProductRequest;
 import com.senior.dto.response.CartResponse;
 import com.senior.dto.response.ProductResponse;
 import com.senior.entities.Cart;
+import com.senior.entities.CartProduct;
 import com.senior.entities.Product;
+import com.senior.repositories.CartProductRepository;
 import com.senior.repositories.CartRepository;
 import com.senior.repositories.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,6 +22,9 @@ public class ProductServiceImpl implements ProductService {
 
     @Autowired
     private ProductRepository productRepository;
+
+    @Autowired
+    private CartProductRepository cartProductRepository;
 
     @Override
     public Product findProductById(String id) {
@@ -46,10 +52,22 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public ProductResponse deleteProductById(ProductRequest request, String id) {
+    public String deleteProductById(ProductRequest request, String id) {
         Product product = request.toEntity();
+        List<CartProduct> list = this.cartProductRepository.getAll();
+        int countHas = 0;
+        for (int i = 0; i < list.size(); i++) {
+           if(list.get(i).getProduct().getId().equals(id)) {
+               countHas++;
+           }
+        }
+        if (countHas > 0) {
+            return "Product already attached to a cart. exclusion Denied.";
+        }
+
         Product updatedProduct = this.productRepository.deleteCartById(product, id);
-        return ProductResponse.fromEntity(updatedProduct);
+        ProductResponse.fromEntity(updatedProduct);
+        return "Product Deleted.";
     }
